Feat: redesign the project architecture with a golang archetype
